requirements.txt
asgiref==3.10.0
black==22.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==3.11.0
django-templated-mail==1.1.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
et-xmlfile==1.1.0
filelock==3.9.0
gunicorn==23.0.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
lxml==4.9.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
numpy==1.23.3
oauthlib==3.2.0
openpyxl==3.0.10
packaging==21.3
pandas==1.5.0
pathspec==0.9.0
Pillow==9.2.0
pipenv==2023.2.4
platformdirs==2.5.2
psycopg==3.1.12
psycopg2-binary==2.9.11
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0.3
requests==2.27.1
requests-oauthlib==1.3.1
rest-api-payload==0.0.5
ruamel.yaml==0.17.35
ruamel.yaml.clib==0.2.8
six==1.16.0
social-auth-core==4.2.0
soupsieve==2.4
sqlparse==0.5.3
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.17.1
virtualenv-clone==0.5.7
whitenoise==6.5.0
